requirements.txt
tensorflow-gpu==1.13.1
numpy==1.16.2
pillow==6.0.0